requirements.txt
aws-secretsmanager-caching==1.1.1.5
botocore==1.32.3
jmespath==1.0.1
psycopg2-binary
psycopg
python-dateutil==2.8.2
six==1.16.0
requests
langchain
PyPDF2==3.0.1
python-dotenv
streamlit
openai
faiss-cpu==1.7.4
altair<5
tiktoken
pgvector
huggingface-hub
InstructorEmbedding
sentence-transformers
pandas
numpy
tiktoken
ipython
beautifulsoup4
playwright
doctran
python-pptx
lxml
pillow
XlsxWriter
unstructured
langchain_experimental
pypdf
openpyxl
langchain_googledrive
nltk
spacy
docx2txt
chardet
pydantic
pandasai
youtube-transcript-api
pytube
GitPython==3.1.31
sqlalchemy
flask-sqlalchemy
load_dotenv
torch>=2.1.0
